Feature: Add navbar style and routes
